Refactor hide purchased

+++ Application/pages/CurrentList.js
@@ -14,6 +14,7 @@ import {
    TopNavigationAction,
    Select,
    Text,
+   CheckBox
 } from 'react-native-ui-kitten';
 import { MenuOutline, AddIcon, BellIcon } from "../assets/icons/icons.js";
 import DoubleClick from "react-native-double-tap";
@@ -36,6 +37,7 @@ const FASTEST_PATH = "FASTEST_PATH";
 const BY_LOCATION = "BY_LOCATION";
 const ORDER_ADDED = "ORDER_ADDED";
 const ALPHABETICALLY = "ALPHABETICALLY";
+const PURCHASED = "PURCHASED";
 
 class CurrentList extends Component {
    constructor(props) {
@@ -71,7 +73,9 @@ class CurrentList extends Component {
 
          minPrice: 0,
          maxPrice: 0,
-         numUnknownPrice: 0
+         numUnknownPrice: 0,
+
+         hidePurchased: false
       };
    }
 
@@ -337,7 +341,9 @@ class CurrentList extends Component {
       }
 
       if (!this.state.firstLoadComplete) {
-         var temp = this.reorganizeListAdded(localItems, localIds);
+         var temp = this.localSort(this.state.orgMethod.value,
+                                   localItems,
+                                   localIds);
 
          localItems = temp.items;
          localIds = temp.ids;
@@ -393,17 +399,19 @@ class CurrentList extends Component {
     * @returns None
     */
    GenerateListItem(item, index) {// Pass more paremeters here...
-      return (
-         <ListItemContainer
-            title={this.getDispName(item)}
-            fromItemView={true}
-            description={item.price === undefined ? "" : "Price: " + item.price.minPrice + " - " + item.price.maxPrice}
-            purchased={item.purchased}
-            listID={this.state.listId}
-            itemID={this.state.listItemIds[index]}
-            onDelete={this.deleteItem}
-         />
-      );
+      if ((!item.purchased) || (item.purchased && !this.state.hidePurchased)){
+         return (
+            <ListItemContainer
+               title={this.getDispName(item)}
+               fromItemView={true}
+               description={item.price === undefined ? "" : "Price: " + item.price.minPrice + " - " + item.price.maxPrice}
+               purchased={item.purchased}
+               listID={this.state.listId}
+               itemID={this.state.listItemIds[index]}
+               onDelete={this.deleteItem}
+            />
+         );
+      }
    }
 
    /**
@@ -473,10 +481,13 @@ class CurrentList extends Component {
       // Call the corresponding selection function
       switch (selectionVal) {
          case ORDER_ADDED:
-            this.reorganizeListAdded();
+            this.localSort(selectionVal);
             break;
          case ALPHABETICALLY:
-            this.reorganizeListAlphabetically();
+            this.localSort(selectionVal);
+            break;
+         case PURCHASED:
+            this.localSort(selectionVal);
             break;
          case BY_LOCATION:
             this.props.navigation.navigate("SelectStorePage", {
@@ -503,16 +514,73 @@ class CurrentList extends Component {
    }
 
    /**
-    * reorganizeListAdded
+    * localSort
     * 
-    * Reorganizes the list to match the order that
-    * the items were added to the list.
+    * Reorganizes the list saved in the state locally,
+    * without needing any external firebase calls.
+    * Uses the sort function designated through selectionVal
+    * to sort the list. Additional parameters can be given
+    * through initItems and initIds to prevent a change of
+    * state.
     * 
-    * @param   None
+    * @param   {String} selectionVal   The type of sort to use
+    * @param   {Array}  initItems   The array of items to sort, if none
+    *                               is given, the list in the state is used
+    * @param   {Array}  initIds The array of ids to sort, if none
+    *                           is given, the list in the state is used
     * 
-    * @returns None
+    * @returns An object containing the sorted items and ids
     */
-   reorganizeListAdded(initItems = null, initIds = null) {
+   localSort(selectionVal, initItems = null, initIds = null){
+      /**
+       * Reorganizes the list to put the items in alphabetical
+       * order based on the item's names.
+       */      
+      function alphSort(a, b) {
+         var itemA = that.getDispName(a.item).toUpperCase();
+         var itemB = that.getDispName(b.item).toUpperCase();
+         return (itemA < itemB) ? -1 : (itemA > itemB) ? 1 : 0;
+      };
+      
+      /**
+       * Reorganizes the list to put the unpurchased items first
+       * and the purchased items last
+       */     
+      function purchasedSort(a, b) {
+         var itemA = a.item.purchased ? 1 : 0;
+         var itemB = b.item.purchased ? 1 : 0;
+         return (itemA < itemB) ? -1 : (itemA > itemB) ? 1 : 0;
+      };
+
+      /**
+       * Reorganizes the list to put it in the order that the
+       * items were added to the list.
+       */   
+      function addedSort(a, b) {
+         var itemA = new Date(a.item.dateAdded);
+         var itemB = new Date(b.item.dateAdded);
+
+         return (itemA < itemB) ? -1 : (itemA > itemB) ? 1 : 0;
+      }
+
+      sortFunction = null;
+      
+      // Retrieve the corresponding selection function
+      switch (selectionVal) {
+         case "ORDER_ADDED":
+            sortFunction = addedSort;
+            break;
+         case "ALPHABETICALLY":
+            sortFunction = alphSort;
+            break;
+         case "PURCHASED":
+            sortFunction = purchasedSort;
+            break;
+         default:
+            break;
+      }
+
+      // Get the items and ids
       var items = [];
       var ids = [];
 
@@ -534,12 +602,7 @@ class CurrentList extends Component {
 
       // Rearrage the nested list to put it in alphabetical order
       var that = this;
-      temp.sort(function (a, b) {
-         var itemA = new Date(a.item.dateAdded);
-         var itemB = new Date(b.item.dateAdded);
-
-         return (itemA < itemB) ? -1 : (itemA > itemB) ? 1 : 0;
-      });
+      temp.sort(sortFunction);
 
       // Retrieve the organized items and ids
       for (var k = 0; k < temp.length; k++) {
@@ -556,47 +619,7 @@ class CurrentList extends Component {
          items: items,
          ids: ids
       }
-   }
 
-   /**
-    * reorganizeListAlphabetically
-    * 
-    * Reorganizes the list to put the items in alphabetical
-    * order based on the item's names.
-    * 
-    * @params  None
-    * 
-    * @returns None
-    */
-   reorganizeListAlphabetically() {
-      console.log('reorganizeListAlphabetically');
-
-      // Get the items and ids
-      var items = this.state.listItems;
-      var ids = this.state.listItemIds;
-
-      // Put the items and their ids in a nested list
-      var temp = [];
-      for (var j = 0; j < items.length; j++) {
-         temp.push({ "item": items[j], "id": ids[j] });
-      }
-
-      // Rearrage the nested list to put it in alphabetical order
-      var that = this;
-      temp.sort(function (a, b) {
-         var itemA = that.getDispName(a.item).toUpperCase();
-         var itemB = that.getDispName(b.item).toUpperCase();
-         return (itemA < itemB) ? -1 : (itemA > itemB) ? 1 : 0;
-      });
-
-      // Retrieve the organized items and ids
-      for (var k = 0; k < temp.length; k++) {
-         items[k] = temp[k].item;
-         ids[k] = temp[k].id;
-      }
-
-      // Update the list state to the reorganized values
-      this.updateListState(items, ids, reorg = true);
    }
 
    /**
@@ -798,14 +821,30 @@ class CurrentList extends Component {
          />
       );
 
-      const renderRightControls = () => [
-         <NotificationAction />,
-         <AddAction />,
-      ];
+      const HidePurchasedAction = (props) => (
+         <CheckBox
+            {...props}
+            checked={this.state.hidePurchased}
+            onChange={(isChecked) => {
+               this._isMounted && this.setState({
+                  hidePurchased : isChecked
+               })
+            }}
+         />
+      )
 
-      const renderRightControl = () => [
-         <AddAction />,
-      ];
+      const renderRightControls = (showBell = false) => {
+         var rightControls = [];
+
+         if (showBell) {
+            rightControls.push(<NotificationAction />);
+         }
+
+         rightControls.push(<HidePurchasedAction />);
+         rightControls.push(<AddAction />);
+
+         return(rightControls);
+      }
 
       const renderMenuAction = () => (
          <TopNavigationAction
@@ -820,7 +859,7 @@ class CurrentList extends Component {
                title={(this.state.listName != "") ? this.state.listName : PAGE_TITLE}
                alignment="center"
                leftControl={renderMenuAction()}
-               rightControls={this.state.userCount > 1 ? renderRightControls() : renderRightControl()}
+               rightControls={renderRightControls(showBell = this.state.userCount > 1)}
             />
             <Layout style={styles.ListContainer}>
                <KeyboardAvoidingView style={styles.container} behavior="position" enabled>
